Separate humanplayer and autoplayer into separate files
